pyproject.toml
[project]
name = "osha-app"
version = "0.1.0"
description = ""
authors = [
    {name = "thomascl24",email = "thomascl@berkeley.edu"}
]
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "langchain-community (==0.3.26)",
    "qdrant-client (>=1.14.3,<2.0.0)",
    "transformers (>=4.53.0,<5.0.0)",
    "datasets (>=3.6.0,<4.0.0)",
    "lightning (>=2.5.2,<3.0.0)",
    "torchmetrics (>=1.7.3,<2.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "sentence-transformers (==4.1.0)",
    "fastapi[standard] (>=0.115.6,<0.116.0)",
    "fastapi-cache2[redis] (==0.1.9)"
]

[tool.poetry]
name = "osha-app"
version = "0.1.0"
description = ""
authors = ["thomascl24 <thomascl@berkeley.edu>"]
readme = "README.md"
packages = [{ include = "osha_app" }]

[tool.poetry.dependencies]
python = "==3.11.13"
pandas = ">=2.3.0,<3.0.0"
langchain-community = "0.3.26"
qdrant-client = ">=1.14.3,<2.0.0"
transformers = ">=4.53.0,<5.0.0"
datasets = ">=3.6.0,<4.0.0"
lightning = ">=2.5.2,<3.0.0"
torchmetrics = ">=1.7.3,<2.0.0"
torch = ">=2.7.1,<3.0.0"
requests = ">=2.32.4,<3.0.0"
pathlib = ">=1.0.1,<2.0.0"
numpy = ">=2.3.1,<3.0.0"
tqdm = ">=4.67.1,<5.0.0"
sentence-transformers = "4.1.0"
fastapi = {version = ">=0.115.6,<0.116.0", extras = ["standard"]}
"fastapi-cache2[redis]" = "0.1.9"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"